Implement Day 1 in Kotlin
New year, new me! Or so they say. In reality, a lot has stayed the same, but a lot is different. I realized that last year, I started to write more detailed explanations of how each day of advent coding went for me as Advent of Code neared its end. I think I want to try that again this year.

It's funny, looking back at my AOC 2023 Day 25 comments, I predicted that I would most likely be using Kotlin again this year, with great success. I think it's also funny how I'm writing these summaries to, realistically, nobody except myself. (If you, dear reader, happen to have found yourself here and reading this, please, take a cookie for your troubles. 🍪) While I put a good deal of effort into my AOC repositories, the code never sees the light of day. It is resigned to drift in the sea of endless content online, waiting to be consumed by web scrapers and fed into AI training datasets. That isn't to say it *matters* whether my code sees the light of day! I do Advent of Code for personal enjoyment; it's also become a bit of a tradition for me at this point. (6 years strong!) It's one of the few traditions I can keep going, in all honestly. Each year, I am surprised to see myself back here on November 30th, both for the speed at which time flies and the fact that I have decided to come back for another round. Yet, traditions aside, it also serves as a snapshot into a sliver of my life every December that I can look back on and compare to see how I've grown (or not). This year, I really let AOC creep up on me; I didn't start setting up my repository until an hour before Day 1 opened!

I did not place today. I think this year, I need to take a bit of pressure off of myself when it comes to placing. It's ok not to place! Quite honestly, speed coding is not my specialty. I tend to enjoy solving problems for the art of solving, not the speed at which I get there. Sometimes, my brain also quite literally cannot think that fast. In such human-like fashion, I tend to compare myself to those who place in the upper echelons of the global leaderboards and wonder why I can't do it as good as them. But this is not a fair comparison to make, and I have to keep that in mind. It's ok to do my best, whatever that may be. Also, my responsibilities only seem to grow with every passing year. I have to remind myself that while Advent of Code is a great passion and hobby of mine, it can't be my whole life, and I only have so much time and energy I can devote without harming my other responsibilities.

While I might have had time tonight to sit down and work on the first day of Advent of Code the moment it opened, I need to remember that this is a luxury and not a given. Time will tell whether I keep up with this repository or let it fall to the side. For my own benefit, I really do hope I can keep up this tradition, and I plan to! What is up in the air (in my mind) is whether I will be able to truly work on AOC this month as it comes out or be forced to wait for freer days. Let this message serve as high, but realistic, expectations for myself this year. I'd like to place some this year, but I also want to be content with just completing all the puzzles and keeping the streak alive. Last year, I (purportedly, according to myself in a past summary) solved ~all puzzles using my own intuition, although some definitely took a good amount of time to reason about and lots of tinkering with solutions. Now, this exceedingly long "summary" is nearing its end, and I thank you for reading, whether that be future me or a weary traveler.