Implement Day 2 in Kotlin
Today was a bit shaky because the servers were not working properly and gave me a lot of Internal Server Errors. I had to open the prose for Day 2 on my phone for it to load. I also got errors while trying to automatically download the input, so I had to rerun my downloader script multiple times until the servers cooperated. I had similar issues while trying to submit in the browser.

I did not place today. I presume that the leaderboards for today will be canceled due to the servers dropping requests.

I had enough time today to again do AOC at puzzle open. We will see if I am this fortunate on future days.