Implement Day 3 in Kotlin
Today was a bit annoying for me, personally. The moment I saw the problem, I knew it would be perfect for a language like Python just because of the versatility of eval().

I decided that Part 1 was simple enough that I would preprocess the input using regex in my IDE. After converting the input to a bunch of lines of xxx*yyy+, I first tried to paste the input into Wolfram Alpha to compute, but it was too long. Next, I tried pasting it as an extremely long expression into the Kotlin code. The compiler generated stack overflow errors when attempting to run as the math expression was just too long. I then tried to split it up into two variables, and crashed the compiler and then froze my IDE. (Curse you, IntelliJ!)

By this point, I had to restart IntelliJ and was a bit exasperated. I decided to just write a quick parser to use my preprocessed input and compute the sum, thus getting Part 1 after some troubles.

For Part 2, I continued to have issues. I decided to write a complex while loop with a lot of mutable variables tracking the current index and finding the minimum of do(), don't(), and mul() and then executing whichever, moving the current index each time. This approach led to multiple bugs, which I eventually fixed. I had to take a break in the middle of working on Part 2 to go perform other responsibilities after about 20 minutes of trying. I came back after ~25 minutes and finished Part 2 at 50 minutes past puzzle open.

Not to be continually negative, but I realized after the fact that by simply opening up a Python REPL with the command "py", I could have pasted in my preprocessed input (excluding newlines) and immediately evaluated the answer to Part 1 without issue. In the moment of solving, I didn't think of this option. I also could've slowed down to stop and think of a smarter solution to Part 2 rather than the complicated solution I landed upon.

Either way, I did not place today. But that's ok! When I place, it tends to be on puzzles that require thinking of the right strategy and put less emphasis on coding quickly. On days when the strategy is obvious or straightforward, I can't really compete because it just comes down to the speed at which you can code up the solution. While Kotlin is concise (in some regards), it is still a compiled and typed language. It doesn't have the same flexibility as Python. I also am rusty at Kotlin, so that doesn't help! All in all, I wasn't expecting to place today, especially after I saw the puzzle for the first time.