Implement Day 3 in Java
This solution is "optimized" to only use a single for loop and no function calls like indexOf or startsWith. Since only basic control flow and elementary types like ints, booleans, and strings are used, this algorithm could be translated into a lower-level language like C with ease.

The only method call is in parse() to turn the list of lines into a single string. This is only necessary due to the way my framework automatically parses the input into a list of lines. In a language like C, one could simply read the file as a single string to begin with. The inclusion of newlines in the data string does not affect this solution, assuming that the operations are not broken across lines (which they are not in my input).

On my hardware, I am benchmarking ~44 us for evaluate().