Implement Day 5 in Kotlin
I wasn't sure how to approach this one in code. I first tried a topological sort, but that wasn't working because I didn't read the problem properly and was including all the ordering rules when I was only supposed to include the rules that had pages contained in the update on a given line. I only realized my misunderstanding of the puzzle after switching to a DFS approach.

For my DFS approach, I realized that I could generate a map of a page to the pages that msut come before it, where an entry in the multimap of [x,y] means that y must come before x at some point. Then, by inspecting the generated map, I saw that I could sort in descending order by the size of the set of each key-value pair and extract the keys so that the first key would be the one that had the most entries in its set, etc. This led me to the expected rule, which I compared against the real rule to see if it matched, and calculated the sum as required. For Part 2, because of the way I designed my solution, it meant that I really only had to change about 2 lines of code to get what I wanted. Rather than checking the expected and actual rules matched, I checked they didn't match, and then used the expected rule to calculate the middle page. This led me to jump from #2464 on Part 1 to #906 on Part 2.

Overall, today was fun. It took a bit of brain power and trial and error to realize what I actually wanted to do (after also realizing what the puzzle was actually asking).