Implement Day 6 in Kotlin
I placed today! #230 on Part 1 and #74 on Part 2, giving me 27 points on the global leaderboard. Today was nothing too crazy. My grid util definitely helped on this one, but I added more util to hopefully aid myself better in the future (the only problem becomes trying to remember what utilities I have in the moment).

For Part 1, I did nothing special, just coded up a solution that simulated what was asked for (albeit I was slow to code this up).

For Part 2, I quickly checked the dimensions of the grid and saw that it was 130x130. This means there are 16,900 locations on the grid, and I knew I could write a brute force solution to just try every single open space as a potential obstacle to create a cycle and have it finish executing in a reasonable amount of time. I slightly modified my code from Part 1 to instead store a pair of coordinate to direction to check for cycles, and the rest was mostly straightforward. I was confident enough in my code to submit my answer without checking against the input, which worked in my favor!

The code took 4.5 seconds to run on my hardware. I think this is one of the few cases where a compiled language like Kotlin (at least Kotlin JVM) shines, as the speedup seems to be really vital here. From what I saw from other people, their brute forces took significantly longer. But maybe this is just from their solution being less optimized, and an interpreted language like Python would do just fine with a solution equivalent to mine. I don't have the time or energy right now to test this out, though.

Overall, I'm really happy to have placed on Day 6, even if it was lower down on the leaderboard. Maybe this year is my year?! Or maybe I'm getting ahead of myself (the likelier option)...