Implement Day 10 in Kotlin
I did not place today. Today was not too bad, very similar to the idea of puzzles in other years (can't recall which). I was surprised at how fast the leaderboard filled up today. With 3 minutes and 36 seconds for Part 1, I placed #183 and was expecting to place higher to be honest.

Part 2 stumped me for a little while until I realized that the magic operation was addition, not multiplication (to be fair, my brain was primed for combinations due to unrelated reasons). After realizing this and some other small tweaks, my original idea for Part 2 worked, and after 15 minutes and 20 seconds, I placed #1305 for Part 2. Quite the "easy" puzzle today, as least for some people!