Implement Day 14 in Kotlin
I did not place today. Part 1 was not bad, but I went slow and had to think about how to do the quadrant calculations. Part 2 was very cute. The key to Part 2 for me was reading "most of the robots should arrange themselves." I assumed the Christmas tree would be a solid shape, so I wrote a flood fill that would check if the number of found robots to robots remaining was greater than a certain percentage and if so, print the grid.

This crude approach was enough to get me the answer to Part 2 (after one accidental wrong answer). My assumptions to Part 2 definitely helped me get the answer, although I was not fast at it. Even though I took 9 minutes from Part 1 to Part 2, I jumped up from #737 on Part 1 to #196 on Part 2. It seems like people are having a hide time finding the tree. I'll take it tonight!

I got stumped on Day 13 Part 2 and haven't solved it yet, so I have not uploaded Day 13's code as it is nonfunctional.