Implement Day 15 in Kotlin
I did not place today. I liked today's puzzle! Part 1 was not too bad, but it made me realize I wanted more utilities for grids. For Part 2, I had the right idea, I just couldn't type it up fast enough to place on the leaderboard. I ended up taking a break and finished it a few hours later. I had 2 small bugs in Part 2 that took a bit of debugging to find.

My first bug was still adding the next position over to the queue even if it was an empty square. My second bug was failing to add the next position over to the queue if a vertical move was being performed as I would only add the counterpart of the box to the queue. After solving these 2 bugs, my code worked for Part 2.

I spent a bit of time writing up more grid utils for the various basic things like looping, mapping, filtering, and summing. At the size of my utilities, I think it would make sense at some point to make a crash course Markdown file or something that I can reference back to in order to see what all utilities I've made and what they do, as I've made utilities over so many years at this point that I forget specific ones until I rediscover them. I think this cheatsheet file should also contain keywords related to each function / utility / category that make it easier to search for and see if what I need already exists in my codebase on the fly. Maybe a task for future me!

Day 13 Part 2 is still stumping me at this point. I haven't worked on it as much as of late. I hope to be able to complete it before Day 25 so that I can get all 50 stars, but we will see.