Implement Day 17 in Kotlin
I did not place today. Today's puzzle was fun, it reminded me slightly of the Synacor Challenge, or really any puzzle from prior years where you have to "optimize" the interpreted code to have any chance of finding the real answer.

For Part 1, it was relatively simple, just implement the operands as required. For Part 2, it took me a bit of thinking. First, I annotated the program to see what it was doing, then I turned it into raw Kotlin and tried finding the right A value that way. While that ran, I then looked at it and thought I could turn it into a program that an SMT solver like Z3 could solve. I let that approach run as I tried to figure out more about the program.

I ended up realizing that the program looks at 3-bit windows of the A value at a time before moving on to the next 3 more significant bits, starting from the least significant bits. It also considers the bits of A that are a few bits higher depending on the current value of the window, and does some math with these values to make an output value. Another byproduct of the 3-bit windows meant that I knew the A value had to be 48 bits long since there were 16 numbers in my program that needed to be output. After drawing out the algorithm and looking at it, I realized I could work backwards from the most significant bits in order to craft a valid A value since the bits to the left of the MSB would always be 0's. By shifting the current A value 3 bits to the left then trying every combination of 3 new bits, you can find which combinations get you the next output number you want, starting from the last number in the program and working backwards.

I ended up using a BFS to find all possible A values (there are more than one!) with this working backwards approach. Finally, taking the minimum of all the valid A values got me the right answer to Part 2! Fun puzzle, definitely took something thinking, but it wasn't insanely out of reach.

As for Day 13 Part 2, I still haven't even looked back at that one yet. I might wait for when I have less responsibilities.