Implement Day 20 in Kotlin
I did not place today. Today was hard for me. I did Part 1 in a really weird way and tried to make it work for Part 2, but it did not. I ended up running some code for Part 2 for 90 minutes only to get a wrong answer. I came back to the puzzle 2 days later and started from scratch. I had a few revelations that helped me solve Part 2, after a lot of trial and error. First, I realized that the shape of the possible cheat end locations around a start position takes a diamond shape that is limited by the Manhattan distance being 20 or less. Then, I noticed that I could simplify the cheat checking by creating 2 maps, a map of the minimum distance from the end to every point in the grid and a map of the minimum distance from the start to every point in the grid. By checking every cheat, I could just look up the distance from the start to that cheat's start position, add it to the distance from the cheat's end position, and add the Manhattan distance between the cheat's start and end positions in order to get the minimum length of the path for that specific cheat.

I ended up having a lot of bugs with over- or under-counting valid/invalid cheats but eventually got the right answer.