Implement Day 21 in Kotlin
Today's puzzle was fun!! It was really a mind-bender to wrap my head around. I had to ponder for a while how I wanted to implement a solution in code for Part 1. My first attempt began with generating the set of all possible path strings between two points at a given depth. One optimization I noticed during this time was that the paths between buttons are "simple", i.e., you only ever take moves that put you closer towards the target. This meant I could look at the change in x and y between the target and current coordinates and find what directions to try. However, even with this, I didn't really know where to take it from here.

By thinking about the puzzle on and off over the course of a few days, I realized I could go for a more directly recursive approach. Instead of explicitly generating each possible path, I would branch recursively on each possible move and try everything. This approach means I have to start at a depth of 0 (the numeric keypad) and recurse down to the depth corresponding to directional keypad that I control, expanding the path taken at each depth. Using the optimization I discovered earlier, if, for example, the target was down and to the right, I would only take moves that put me down or to the right. Each time I take a move, I record it as part of the path I've taken. Once I reach the target button, I then recurse one depth further down with that path as the new code (plus A to enter). Once I reach the final depth (the directional keypad controlled by me), the length of the shortest sequence is simply the length of the path since I can press any amount of buttons in any order, with only one step required for each. I then return the length of the shortest sequences between left/right or up/down.

I had a little of trouble with my recursive solution at first, because I thought I would simply return 1 when I reached the final depth, but I actually needed to return the length of the code provided at that depth. Originally, I also added on the length of the shortest sequence of the remaining characters in the code after finding the minimum between left/right and up/down, but I realized I needed to instead calculate this length after reaching the target corresponding to the current character. These 2 fixes were sufficient for me to solve Part 1!

For Part 2, it was a simple optimization problem, and it didn't take much effort for me compared to Part 1. I was lucky in that my recursive solution didn't actually track the sequence string itself, only the length of the shortest sequence. So it was easy enough to generalize to go up to a maximum depth of 26 instead of 3. To solve, I first realized that it would probably be good to memoize the recursive function. Repeat states (states where a robot at the same depth is going between the same start and target coordinates with the same code) are bound to happen because of the small number of possible states. This memoization was enough to make my solution work for Part 2.

Overall, while I didn't place on today's puzzle (I didn't even open it until a day after!), I really enjoyed the puzzle as it required using my brain, and I felt very satisfied after solving it. It was really nice to be able to solve Part 2 so easily, and I didn't have any wrong answers for this puzzle.