Implement Day 24 in Kotlin
Today's puzzle was very fun. Part 1 was not too bad, but Part 2 took me a while. I first noticed that the input numbers were 45 bits long, which meant that iterating through every combination of inputs would have been much too inefficient (and also would not really have done anything).

I then tried to identify which bit places were faulty by calculating 1+0, 0+1, 1+1, and 0+0 at each bit place. However, this gave me too many gates to check to simply try swapping 4 pairs at random and validate if the system was corrected. Once I realized this strategy wasn't going to work, I decided to go to bed and try again later.

After sleeping on it, I realized I wanted to use Part 2 as an opportunity to try to design my own binary adder, without looking at my input for inspiration. I thought if I could design my own binary adder (and assuming it matched up with how the input's binary adder worked), I could work backwards by programmatically generating that system and comparing to my input, then matching up gates and seeing the discrepancies. (I think this solution would still work, but it's not the one I went for.)

For designing my binary adder, it was a nice refresher on logic gates and such. My first attempt at a design was bad. My second design actually was mostly correct, at least compared to the input. My one oversight was that for an input of a=1, b=1, carry=1, XOR of a with b, and XOR of that with the carry bit, outputs 1 which is actually what you want, since 1+1+1 should put out a 1 for that bit place. In my second design, I had 2 AND gates between a, b, and the carry in order to achieve this same effect, and OR'd into the output bit. I realized this redundancy by looking at the input, and corrected it for my final design. Using this final design, I then wrote some code that would check the gates and wires for each bit place and see if it matched what I expected. I didn't know how to actually write code that would uncross the wires, so I just errored out with what wire was crossed once I found it. I manually fixed the crossed pairs one by one while noting down which ones I swapped. Since there are only 4 pairs of crossed wires, it was easy enough to do by hand with the helper code I wrote. This got me the star for Part 2.

For actually writing a solver for Part 2, I realized that Eric was nicer in the input than he said in the puzzle prose. For my input, the crossed wires appeared to only swap within the binary adder for a single bit place, and there was only at most one pair of crossed wires for a given bit place's binary adder. I decided to use these assumptions to craft my solution, as it makes the code significantly easier.

I tested my solution on 3 inputs and got the correct answer for both parts on all 3, so I believe my assumptions probably work for the inputs that Eric crafted.

Overall, this puzzle was very fun and was a nice reason to try my hand at creating my own binary adder.