Implement Day 13 in Kotlin
After a year, I've finally solved it! Originally, I solved Part 1 on the night it released but got stumped by Part 2 (the classic "now figure out the actual math behind this rather than brute force" twist). After a few more failed attempts at solving it sporadically, I got busy with life and never got back around to trying to solve it.

However, with AOC 2025 coming up (although in half the length of prior years, sad but understandable! Thanks for all your work over these 10 years, topaz!), I felt obligated to try to finish out AOC 2024 in preparation. It took me a lot of thinking and mathematical derivations, calling back to my Number Theory coursework, in order to be able to solve this problem.

The first major roadblock was coming up with a feasible solution. I kept oscillating back and forth between trying to apply the Chinese remainder theorem and doing something more complicated with LCM/GCD; obviously, I ended up doing the more complicated solution. (In hindsight, I don't think CRT makes sense to apply here, but it *felt* like it was relevant, at least to my brain.)

One thing that made the problem easier is that the math is the exact same for Part 1 and Part 2, save for the actual offset to the prize location. After deriving the math and seeing it apply successfully to the two claws in the example from Part 1, I then moved to getting the math to work out on my actual input for Part 1. Unfortunately, I couldn't find my original brute force solution to Part 1 (I never committed my half-completed solution, and it's been a year since), so I didn't have a baseline to compare against.

For my solution, what I ended up doing was realizing that a form of Bézout's identity would be immeasurably helpful here, for the X and Y axes separately. Specifically, the two numbers given by the values of the A and B button for a specific axis combined with the location of the prize on that axis form the basis of the identity. An implementation of the extended Euclidean algorithm while additionally tracking the coefficients of the identity was trivial. I knew that the solution to an individual axis would be an integer multiple of Bézout's identity for that axis, one that makes the GCD equal to the location of the prize (i.e., the GCD of the buttons on that axis must divide the prize location nicely as a necessary-but-not-sufficient condition for a solution to be possible). Finally, I used the idea that any feasible solution could be represented in the form of this integer-multiple of Bézout's identity with an additional term equally added/subtracted within the two coefficients to keep everything equal. The values of these coefficients, when solved correctly, lead directly to the minimum number of A presses and B presses. Specifically, one must solve the system of linear equations given by matching up these coefficients from the X and Y axis along the A and B buttons (as seen in the comments of my solution).

The remaining hurdles ended up being errors in my mathematical derivation and not filtering out enough of the impossible machine configurations. It took me a good deal of effort to figure out that if my derived variable t was non-integral, it meant there was no integer combination of A and B presses that would land on the prize (i.e., it is impossible). Obviously, looking back on it, this is because with the baseline solution given by the integer multiple of Bézout's identity, we can only "exchange" A and B presses at an integral rate. After properly filtering out invalid configurations, my answer to Part 1 finally matched my original answer that I submitted to the website almost a year ago. Switching the code to use Java's BigInteger class (largely for peace of mind) and adding in the offset for Part 2's prize locations directly led me to getting the correct answer to Part 2!

Overall, this puzzle was quite fun to reason about, even if it did take a year (on and off; mostly off) to do. I presume that my solution is probably overcomplicated to a large degree, but it's mine, and that's what matters! I still have Day 25 left to do, which I largely gave up on since I didn't have Day 13 done by the time that that day's challenge released.