Simplify Day 13 in Kotlin
Gosh, after taking a glance at the subreddit for Day 13 solutions, I sure do feel silly with the extreme degree to which I overcomplicated the problem! Funny how I landed on solving a system of linear equations but added so much unnecessary cruft in the middle.

After simplifying my solution to derive the math and solve for the much-cleaner system of linear equations, it worked first try! Well, at least I can rest happy knowing I've solved Day 13 in a not-terribly-convoluted way.