Implement Day 25 in Kotlin
And with that, it's done! Day 25 was suspiciously easy; I'm sure I've felt this way in prior years, but it feels like topaz makes Day 25 puzzles intentionally easier than the otherwise harder-difficulty puzzles of the last few puzzles of the year, to make for a nice ending as people prepare for Christmas festivities the following day.

By quite literally following the conventions laid out in the prose to convert the key/lock grids to their respective heights, a simple product match while ensuring all columns sum to be less than or equal to 5 was enough to count up all the potential matching key/lock pairs and solve Part 1 without much fuss.

As is tradition, there was no Part 2 challenge for today. I enjoyed reading the ending; I still have to go back through all the puzzles and look at the easter eggs left behind! I have the added challenge of not having touched any of these puzzles since a year ago, so I have to remember what is even being talked about. Funny how my memory leads to me almost entirely forget about these puzzles and their contents. (In the best case, I only spend a single night on each of them, so I feel like it's *somewhat* justified.)

From what I can recall, I enjoyed this year's set of puzzles, and while I am excited for AOC 2025 (it's the entire reason I felt motivated to finish AOC 2024), I trek on with the solemn knowledge that it will be half the length from now on. No longer will I have a month-long ritual lasting up to Christmas Eve, but merely Christmas half-Eve, so to speak. I wonder how the difficulty scaling will work out this year?

I've been partaking in the Advent of Code challenges since 2018, and I went back and did 2015, 2016, and 2017 in my spare time interspersed in there. Perhaps it is the nature of the beast, but nostalgia is a fickle thing. I've greatly loved each and every year's puzzles, but it *feels* like I had greater fun when I was first starting out. In 2018 and 2019, I was very much a more clueless programmer, often getting stuck with the problems, but still enjoying them to a great deal. I much loved the Intcode saga from 2019, and while I'm sure it's controversial because of its inherit need for you to keep up by crafting working solutions to prior (albeit odd-numbered) days in order to participate, it was so fun to expand upon an entire virtual architecture over the course of a month and see the number of complicated, over-the-top tasks it could take on. (I distinctly remember the jumping space robot being both incredibly cool and frustrating to solve-- I don't think I ever properly solved that one fully unassisted.) I even went and did the Synacor Challenge in my thirst for more AOC-like content; it was quite fun to see the thought process of topaz in an entirely different format that predates AOC. (If, intrepid reader, you find yourself here also thirsting for such an adventure and curious to learn more about the Synacor Challenge, check out an archive of it here: https://github.com/Aneurysm9/vm_challenge ) In 2020, 2021, and 2022, I felt like my skills grew immeasurably year-over-year, and my efforts were rewarded with top-spot placements on the global leaderboards for specific days. However, 2023 and 2024 have felt like a struggle. Maybe it's simply that it feels hard to see the new and intrigue in something when you have done (checks notes) 350 things like it in the past. Or maybe it's because I felt myself slipping, the idea of placing on the leaderboard becoming ever more so out of reach. It's important to remind myself that leaderboard placement has very little to do with this whole thing called being a programmer; also the leaderboard quite literally gets more competitive each year with the natural growth of things. And perhaps it's because I've now entirely switched to Kotlin, a language even higher-level than Java (which is a tough bar to beat). While its syntactic sugar constructs are nice, sometimes it feels like I'm wrangling a beast I don't fully control, unaware of what is truly happening. This definitely is in part because of my relative lack of experience using Kotlin; typically I pick it back up each year with the start of AOC. Perhaps it's time to switch back to good old Java, of which I know the ins and outs? While it may be more typing, at least I can more easily reason about the logic and flow of things, rather than drowning myself in a snowstorm frenzy of maps, lets, and Elvis operators (who let Elvis Presley into MY Kotlin anyways?!).

All in all, Advent of Code has been quite the roller coaster. It's one of the few traditions I've kept up with these past 6 (coming up on 7) years through all my various stages of life. Especially in the beginning, this was most definitely a time-consuming hobby, and a mind-consuming one. I would often bang my head against a problem after it released at night and have to put a pin in it until the next day, hoping to think up a solution in between the schedule of my busy day. Sometimes, I was successful and would solve the challenge of the prior night before the next one released. Other times, I was not so lucky. It's funny thinking back on how in those years, I would attempt a problem enough times and get frustrated enough to just "cheat" and look up a solution to implement myself. Now, I try to pride myself on coming up with solutions to all the days myself. It's definitely not easy; some subjects are more natural for me than others. But it's the effort, determination, and willpower that make it important to me to try to hold myself to. In some cases, this can backfire on me, and I can get stubborn with myself to solve something. In a way, AOC is a nice microcosm of my brain and how I approach and tackle problems, along with a view of how that process breaks down when I get stuck. I can't begin to fathom how participating in AOC has shaped the programmer I have become today. The transformation is very much a tacit one; I *feel* different from when I started out on this journey. Part of that has been participating in AOC itself and learning things on the go, and part of that has been the knowledge and ideas and strategies I have been able to take away from the process and apply or relate to things I see in my programming work outside of festive-themed coding puzzles. I am very much happy and appreciative for that journey.

Now, I look forward to AOC 2025, whatever its new form ends up looking like! I'm not sure why I've opted to write such long and detailed commit messages for these solves as of late, but I think it's my own way of trying out a very low-stakes variant of blogging, with none of the commitment or extra work required. I know with almost certainty that no one except myself will ever read these long, drawn-out essays, but that is part of the appeal. These "blog posts", if you will, are a sign post to myself about ME and who *I* am in this exact moment. It's a portal into my current headspace, my joys and frustrations with the problem of the day and otherwise. It's one of the few ways I can look back at this thing called "Advent of Code" and remember what at all I was doing. I only wish I would have done something similar since the beginning, to be able to see just how stark of a difference there is between then and now. In any case, I think I will try to keep this trend up, for the foreseeable future. On to AOC 2025!